backend/requirements.txt
# Web Framework
Flask==2.2.5
Flask-CORS>=4.0.0
gunicorn==21.2.0
Werkzeug==2.2.3
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
# AI & RAG
langchain>=0.1.16,<0.2.0
langchain-community>=0.0.25,<0.1.0
langchain-groq==0.1.0
groq>=0.4.1,<1
sentence-transformers==2.2.2
faiss-cpu==1.7.4; platform_system != "Windows"
# Machine Learning
tensorflow-cpu>=2.16.0,<2.17.0
joblib>=1.3.0
mlflow>=2.14.0,<3.0.0

# Data Processing
scikit-learn>=1.3.0,<1.4.0
pandas>=2.1.0,<2.2.0
numpy>=1.24.0,<1.27.0

# Technical Analysis
ta>=0.11.0

# API Client
requests>=2.31.0
aiohttp==3.9.1
yfinance==1.3.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0

# Unified MLOps Dependencies
apache-airflow==2.10.3
great-expectations==1.2.4
evidently==0.6.2
xgboost==2.1.3
optuna==4.1.0
feast==0.42.0
prometheus-client==0.21.1
redis==5.2.1

# Keep TensorFlow runtime compatible with known protobuf constraints.
protobuf>=4.25.3,<5
